GBCompo23: add resource links for GBSDK
Just a suggestion, if people are inclined to add this to the resources section as a tool option.

Note, two of the linked resources are specific game implementations since those are the only two fully implemented examples i know of that exist. For those who might use this tool they could be useful references.